Fixed Microsoft Visual Studio warnings in examples

+++ examples/common/hud.h
@@ -143,7 +143,7 @@ protected:
             v = std::max(std::min(v, max), min);
             if (intStep) {
                 // MSVC 2010 does not have std::round() or roundf()
-                v = v>0.0f ? floor(v+0.5f) : ceil(v-0.5f);
+                v = v>0.0f ? floorf(v+0.5f) : ceilf(v-0.5f);
             }
             value = v;
         }
